Use as much as possible of system libraries
Ideally we should not use Mason at all for a Debian package and
rely on system dependencies that are shared with other packages.

It turns out it was not possible for Ubuntu 15.04 and currently we
are using Mason for:

- glfw3: Ubuntu's version is a bit older than we need and doesn't
         have glfwPostEmptyEvent().

- libpng: Dunno why but Ubuntu has a fairly outdated version of libpng
          (v1.2.51) and we require 1.6.16.

- nunicode: not packaged by Ubuntu.

It would probably be acceptable to ship the source code of these
dependencies together with the Debian package and link statically.